add game loop and control logic
Add a keyboard event listener, and tie it all together with the game loop.  We
select on two channels: the keyboard input and a timeout channel representing a
tick, and generate a direction event on an output channel with every tick: a
new direction, if one came from the keyboard, or the previous direction
otherwise. At each game iteration, we check to see if we've eaten an apple, and,
if so, shorten the game tick to speed up the snake movement.

Signed-off-by: Iliyan Malchev <malchev@google.com>